@@ -2,7 +2,13 @@
   "name": "@masatomakino/gulptask-ejs",
   "version": "0.3.3",
   "description": "ejs compiler task for gulp.js",
-  "main": "./bin/index.js",
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./bin/index.d.ts",
+      "default": "./bin/index.js"
+    }
+  },
   "types": "./bin/index.d.ts",
   "repository": {
     "type": "git",
@@ -18,23 +24,21 @@
   },
   "dependencies": {
     "chokidar": "^3.5.3",
-    "commander": "^10.0.0",
-    "ejs": "^3.1.8",
-    "glob": "^8.1.0"
+    "commander": "^11.1.0",
+    "ejs": "^3.1.9",
+    "glob": "^10.3.10"
   },
   "devDependencies": {
-    "@types/ejs": "^3.1.1",
-    "@types/glob": "^8.0.1",
-    "@types/node": "^18.13.0",
+    "@types/ejs": "^3.1.3",
+    "@types/node": "^20.8.6",
     "husky": "^8.0.3",
     "prettier": "^2.8.4",
     "pretty-quick": "^3.1.3",
-    "typescript": "^4.9.5"
+    "typescript": "^5.2.2"
   },
   "scripts": {
     "build": "tsc",
     "dev": "tsc -w",
-    "run": "node gulpfile.js",
     "testRun": "node bin/CLI.js --srcDir ./testEjs --distDir ./dist --componentPatterns '**/_*.ejs'",
     "buildEJS": "npx @masatomakino/gulptask-ejs --srcDir ./testEjs --distDir ./dist --componentPatterns '**/_*.ejs'",
     "watchEJS": "npm run buildEJS -- -W"
